refactor(parts): move bundle plugin and tests
