+++ charmcraft/parts/__init__.py
@@ -19,8 +19,7 @@
 import os
 import pathlib
 import shlex
-import sys
-from typing import Any, Dict, List, Set
+from typing import Any, Dict, List
 
 from craft_cli import CraftError, emit
 from craft_parts import LifecycleManager, Step, plugins
@@ -29,6 +28,7 @@ from craft_parts.parts import PartSpec
 from xdg import BaseDirectory  # type: ignore
 
 from charmcraft import charm_builder, instrum
+from charmcraft.parts.bundle import BundlePlugin
 from charmcraft.parts.charm import CharmPlugin, CharmPluginProperties
 from charmcraft.reactive_plugin import ReactivePlugin, ReactivePluginProperties
 
@@ -37,69 +37,10 @@ __all__ = [
     "CharmPluginProperties",
     "ReactivePlugin",
     "ReactivePluginProperties",
-    "BundlePluginProperties",
-    "BundlePlugin",
     "setup_parts",
     "process_part_config",
     "PartsLifecycle",
 ]
-
-
-class BundlePluginProperties(plugins.PluginProperties, plugins.PluginModel):
-    """Properties used to pack bundles."""
-
-    source: str
-
-    @classmethod
-    def unmarshal(cls, data: Dict[str, Any]):
-        """Populate bundle properties from the part specification.
-
-        :param data: A dictionary containing part properties.
-
-        :return: The populated plugin properties data object.
-
-        :raise pydantic.ValidationError: If validation fails.
-        """
-        plugin_data = plugins.extract_plugin_properties(
-            data, plugin_name="bundle", required=["source"]
-        )
-        return cls(**plugin_data)
-
-
-class BundlePlugin(plugins.Plugin):
-    """Prepare a bundle for packing.
-
-    Extra files to be included in the bundle payload must be listed under
-    the ``prime`` file filter.
-    """
-
-    properties_class = BundlePluginProperties
-
-    @classmethod
-    def get_build_snaps(cls) -> Set[str]:
-        """Return a set of required snaps to install in the build environment."""
-        return set()
-
-    def get_build_packages(self) -> Set[str]:
-        """Return a set of required packages to install in the build environment."""
-        return set()
-
-    def get_build_environment(self) -> Dict[str, str]:
-        """Return a dictionary with the environment to use in the build step."""
-        return {}
-
-    def get_build_commands(self) -> List[str]:
-        """Return a list of commands to run during the build step."""
-        install_dir = self._part_info.part_install_dir
-        if sys.platform == "linux":
-            cp_cmd = "cp --archive --link --no-dereference"
-        else:
-            cp_cmd = "cp -R -p -P"
-
-        return [
-            f'mkdir -p "{install_dir}"',
-            f'{cp_cmd} * "{install_dir}"',
-        ]
 
 
 def setup_parts():
