+++ tests/conftest.py
@@ -32,6 +32,7 @@ import yaml
 from craft_parts import callbacks, plugins
 from craft_providers import Executor, Provider
 
+import charmcraft.parts
 from charmcraft import deprecations, instrum, parts
 from charmcraft.bases import get_host_as_base
 from charmcraft.models import charmcraft as config_module
@@ -383,6 +384,28 @@ def charm_plugin(tmp_path):
     }
     plugin_properties = parts.CharmPluginProperties.unmarshal(spec)
     part_spec = plugins.extract_part_properties(spec, plugin_name="charm")
+    part = craft_parts.Part(
+        "foo", part_spec, project_dirs=project_dirs, plugin_properties=plugin_properties
+    )
+    project_info = craft_parts.ProjectInfo(
+        application_name="test",
+        project_dirs=project_dirs,
+        cache_dir=tmp_path,
+    )
+    part_info = craft_parts.PartInfo(project_info=project_info, part=part)
+
+    return plugins.get_plugin(part=part, part_info=part_info, properties=plugin_properties)
+
+
+@pytest.fixture()
+def bundle_plugin(tmp_path):
+    project_dirs = craft_parts.ProjectDirs(work_dir=tmp_path)
+    spec = {
+        "plugin": "bundle",
+        "source": str(tmp_path),
+    }
+    plugin_properties = charmcraft.parts.bundle.BundlePluginProperties.unmarshal(spec)
+    part_spec = plugins.extract_part_properties(spec, plugin_name="bundle")
     part = craft_parts.Part(
         "foo", part_spec, project_dirs=project_dirs, plugin_properties=plugin_properties
     )
